+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Adds meta keywords to site.
 - Adds a method for updating the site's beautiful-jekyll copy.
 - Updated the site's theme and added some css tweaks like transitions.
+- Added test to ensure no trailing whitespace and that we do file ending newlines.
 
 ## [v3.16.0](https://github.com/lexicalunit/spellbot/releases/tag/v3.16.0) - 2020-08-07